# gadget-store/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Commerce API",
        "description": "API documentation for Laravel Commerce application with product management, sales tracking, categories, and admin authentication",
        "contact": {
            "email": "admin@laravelcommerce.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/admins": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get all admins",
                "description": "Retrieve a list of all administrators",
                "operationId": "getAdmins",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Admin"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Create a new admin",
                "description": "Create a new administrator account",
                "operationId": "createAdmin",
                "requestBody": {
                    "description": "Admin data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Smith"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.smith@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admin"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admins/{id}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get admin by ID",
                "description": "Retrieve a specific administrator by their ID",
                "operationId": "getAdminById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "68b74ba7002cda59000d800c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admin"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Update admin",
                "description": "Update an existing administrator",
                "operationId": "updateAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "68b74ba7002cda59000d800c"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Admin data to update",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Smith Updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.updated@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "newpassword123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+1987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admin"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Delete admin",
                "description": "Delete an administrator",
                "operationId": "deleteAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "68b74ba7002cda59000d800c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Admin deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin not found"
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Admin login",
                "description": "Authenticate admin with email and password",
                "operationId": "adminLogin",
                "requestBody": {
                    "description": "Login credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "admin": {
                                            "$ref": "#/components/schemas/Admin"
                                        },
                                        "admin_id": {
                                            "type": "string",
                                            "example": "68b74ba7002cda59000d800c"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email or password is incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Admin": {
                "title": "Admin",
                "description": "Admin model",
                "properties": {
                    "id": {
                        "description": "MongoDB-like ObjectId",
                        "type": "string",
                        "example": "68b74ba7002cda59000d800c"
                    },
                    "name": {
                        "description": "Admin name",
                        "type": "string",
                        "example": "John Smith"
                    },
                    "email": {
                        "description": "Admin email address",
                        "type": "string",
                        "format": "email",
                        "example": "john.smith@admin.com"
                    },
                    "phone_number": {
                        "description": "Admin phone number",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-02T20:30:15.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-02T20:30:15.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "AdminAuth": {
                "type": "apiKey",
                "description": "Admin ID for authentication (required for create, update, delete operations)",
                "name": "Admin-ID",
                "in": "header"
            }
        }
    },
...
}